renamed phases to train and pretrain in config
